remove inputs
Due to legality & ethical reasons.
https://adventofcode.com/2022/about#legal

/u/topaz2078: "I don't mind having a few of the inputs posted, please
don't go on a quest to collect many or all of the inputs for every
puzzle. Doing so makes it that much easier for someone to clone and
steal the whole site. I put tons of time and money into Advent of
Code, and the many inputs are one way I prevent people from copying
the content."